refactor(structure): move errors to lib
