+++ .nvmrc
@@ -1,0 +1,1 @@
+v21.2.0
