+++ Makefile
@@ -2,10 +2,11 @@
 .DEFAULT_GOAL := docker
 
 init:
-	clear
-	cd website && npm install
-	cd api && npm install
-	cd website/server && npm install
+	@cd website && npm install
+	@cd api && npm install
+	@cd website/server && npm install
+	@cd website && npm run build
+
 
 npm:
 	clear
