+++ api/lib/errors/http/AuthorizationError.js
