+++ api/lib/errors/http/BadSheetDataError.js
@@ -2,8 +2,8 @@
  * Should be thrown if the data from a sheet is not in the expected format.
  * @deprecated
  */
-export default class BadSheetDataError extends Error{
-    constructor(message){
+export default class BadSheetDataError extends Error {
+    constructor(message) {
         super(message);
         this.name = 'BadSheetDataError';
         this.status = 500;
