+++ api/lib/errors/http/NotFoundError.js
