+++ api/lib/errors/http/StudentNotFoundError.js
