+++ api/lib/errors/redis/Base.js
