+++ api/lib/errors/redis/MisformedKeyError.js
@@ -7,10 +7,10 @@ export default class MisformedKeyError extends Base {
     constructor(message, { expectedType, receivedValue } = {}) {
         let smartMessage = message;
         if (expectedType && receivedValue) {
-            smartMessage += `; expected ${expectedType}, but got ${typeof receivedValue}`
+            smartMessage += `; expected ${expectedType}, but got ${typeof receivedValue}`;
         }
         super(smartMessage);
-        this.name = 'MisformedKeyError'
+        this.name = 'MisformedKeyError';
         this.expectedType = expectedType;
         this.receivedValue = receivedValue;
     }
