+++ api/v2/Routes/isadmin/index.js
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import { isAdmin } from '../../../lib/userlib.mjs';
 import { getEmailFromAuth } from '../../../lib/googleAuthHelper.mjs';
-import AuthorizationError from '../../../lib/HttpErrors/AuthorizationError.js';
+import AuthorizationError from '../../../lib/errors/http/AuthorizationError.js';
 const router = Router({ mergeParams: true });
 
 // Responds with whether or not the current user is an admin
@@ -9,7 +9,7 @@ router.get('/', async (req, res) => {
     try {
         const authHeader = req.headers['authorization'];
         if (!authHeader) {
-            throw new AuthorizationError("Authorization Header is empty.");
+            throw new AuthorizationError('Authorization Header is empty.');
         }
         const authEmail = await getEmailFromAuth(authHeader);
         const adminStatus = await isAdmin(authEmail);
@@ -21,7 +21,9 @@ router.get('/', async (req, res) => {
                 return res.status(401).json({ message: err.message });
             default:
                 console.error('Internal Server Error:', err);
-                return res.status(500).json({ message: 'Internal Server Error' });
+                return res
+                    .status(500)
+                    .json({ message: 'Internal Server Error' });
         }
     }
 });
