+++ website/package.json
@@ -17,7 +17,7 @@
     "react-router-dom": "^6.7.0"
   },
   "scripts": {
-    "start": "cd ../server && npm start",
+    "start": "cd ../server && npm run dev",
     "startCurrentBuild": "nodemon index.js",
     "react": "react-scripts start",
     "build": "react-scripts build && rm -rf server/build && mv build server",
